requirements.txt
boto3==1.39.3
botocore==1.39.3
certifi==2025.6.15
charset-normalizer==3.4.2
DateTime==5.5
distro==1.9.0
greenlet==3.2.3
idna==3.10
jmespath==1.0.1
numpy==2.3.1
pandas==2.3.0
psycopg2-binary==2.9.10
pyaml==25.5.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
s3transfer==0.13.0
setuptools==80.9.0
six==1.17.0
SQLAlchemy==2.0.41
tabula==1.0.5
tabula-py==2.10.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
zope.interface==7.2
